Improved errors. Moved mime type and error to separate file
