+++ Sources/KituraRequest/Encoding/MultipartFormData/BodyPart+MimeType.swift
@@ -1,0 +1,46 @@
+/**
+ * Copyright Sergey Minakov 2016
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ **/
+
+
+import Foundation
+
+extension BodyPart {
+
+    public enum MimeType {
+
+        public enum Image: String {
+            case any = "*"
+            case png = "png"
+            case jpeg = "jpeg"
+        }
+
+        case none
+        case text
+        case image(Image)
+
+        var value: String? {
+            switch self {
+            case .image(let type):
+                return "image/\(type.rawValue)"
+            case .text:
+                return "text/plain"
+            default:
+                break
+            }
+            return nil
+        }
+    }
+}
